First commit
Deck_testing_v1.0 is a modified version of the black and white face card file. Some of the modifications I made were: gave the cards meaningful id names, grouped internal card assets in standard poker card terminology like the "index" of the card, began the process of standardizing the deck assets so they can be reused as defs, determined the proper scaling factor to transition from the count_suit to the index_suit. 
Progess.txt are notes for later use in writing a python script to fully generate the svg file from scratch